feat : 이미지 drag and drop 기능 추가
